geom3d works
