+++ rrtstar/CMakeLists.txt
@@ -112,6 +112,7 @@ catkin_package(
 ## Your package locations should be listed before other locations
 include_directories(
  include
+ include/rrtstar
   ${catkin_INCLUDE_DIRS}
   ${CMAKE_CURRENT_SOURCE_DIR}
 )
